pyproject.toml
[project]
package-mode = false
name = "agentalchemy"
version = "0.1.0"
description = "A platform for building and configuring AI Agents using Langgraph"
authors = [
    {name = "mauro",email = "m.montenegro.meza@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "langgraph (>=0.4.3,<0.5.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-aws (>=0.2.23,<0.3.0)",
    "langchain-tavily (>=0.1.6,<0.2.0)",
    "rich (>=14.0.0,<15.0.0)",
    "newsapi-python (>=0.2.7,<0.3.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "mangum (>=0.19.0,<0.20.0)",
    "feedparser (>=6.0.11,<7.0.0)",
    "googlenewsdecoder (>=0.1.7,<0.2.0)",
    "lxml (>=5.4.0,<6.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "google-cloud-bigquery (>=3.35.1,<4.0.0)",
    "pydantic-settings (>=2.0.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
